+++ src/ccl/app.module.ts
@@ -14,7 +14,7 @@ import { ENV_PROVIDERS } from '../platform/environment';
 import { routing, appRoutingProviders } from './app.routing';
 import { AppComponent } from './app.component';
 import { LabComponent } from './lab/lab.component';
-import { PageNotFoundComponent } from './http-statuses/page-not-found/page-not-found.component';
+import { PageNotFoundComponent } from './http-status/page-not-found/page-not-found.component';
 import { ChartService } from './services/chart.service';
 import { IndicatorsService } from './services/indicators.service';
