+++ src/ccl/app.routing.ts
@@ -2,7 +2,7 @@ import { ModuleWithProviders } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
 import { LabComponent } from './lab/lab.component';
-import { PageNotFoundComponent } from './http-statuses/page-not-found/page-not-found.component';
+import { PageNotFoundComponent } from './http-status/page-not-found/page-not-found.component';
 
 const appRoutes: Routes = [
   { path: '', component: LabComponent },
